[Chore] 치명적 타입 추론 버그 해결

+++ LateForDinner/Assets/Scripts/UserInterfaces/Popups/UIQuestInventoryPopup.cs
@@ -236,7 +236,7 @@ public class UIQuestInventoryPopup : UIPopup, IDraggablePopup, IFocusablePopup
         {
             GetText(Texts.HealthTabText).text = player.Attributes.GetBase<int>(AttributeType.Health).Value.ToString();
             GetText(Texts.MoveSpeedTabText).text = player.Attributes.GetBase<float>(AttributeType.MoveSpeed).Value.ToString("F1");
-            GetText(Texts.DamageTabText).text = player.Attributes.GetBase<float>(AttributeType.Damage).Value.ToString("N0");
+            GetText(Texts.DamageTabText).text = player.Attributes.GetBase<int>(AttributeType.Damage).Value.ToString("N0");
             GetText(Texts.AttackSpeedTabText).text = player.Attributes.GetBase<float>(AttributeType.AttackSpeed).Value.ToString("F2");
             GetText(Texts.JumpForceTabText).text = player.Attributes.GetBase<float>(AttributeType.JumpForce).Value.ToString("F1");
             GetText(Texts.JumpCountTabText).text = player.Attributes.GetBase<int>(AttributeType.JumpCount).Value.ToString();
